Add support for quarter_information field
This isn't fully FireRoad-compatible, since AFAIK there is no way to
programmatically get the start/end dates of half-term subjects from the
Data Warehouse (so this information is left out).

It is almost certain that some fields will eventually need to be filled
in by scraping the current catalog online (this does mean that these
fields cannot be populated for historical subjects); at that stage, we
can add start/end dates of half-term subjects.